update types and support collections and tables
