+++ frontend/src/metabase-types/api/search.ts
@@ -1,4 +1,3 @@
-import { CardDisplayType } from "./card";
 export type SearchModelType =
   | "dashboard"
   | "card"
@@ -6,11 +5,3 @@ export type SearchModelType =
   | "collection"
   | "table"
   | "database";
-
-export type SearchEntity = {
-  id: number;
-  name: string;
-  model: SearchModelType;
-  description?: string;
-  display?: CardDisplayType;
-};
