+++ frontend/src/metabase/visualizations/visualizations/LinkViz/EntityDisplay.tsx
@@ -2,9 +2,9 @@ import React from "react";
 
 import Icon from "metabase/components/Icon";
 import { ItemIcon } from "metabase/search/components/SearchResult";
-import type { SearchEntity } from "metabase-types/api";
 import { color } from "metabase/lib/colors";
 import Ellipsified from "metabase/core/components/Ellipsified";
+import type { LinkEntity } from "./types";
 
 import { EntityDisplayContainer, LeftContainer } from "./EntityDisplay.styled";
 
@@ -12,7 +12,7 @@ export const EntityDisplay = ({
   entity,
   showDescription = false,
 }: {
-  entity: SearchEntity;
+  entity: LinkEntity;
   showDescription?: boolean;
 }) => {
   return (
