+++ frontend/src/metabase/visualizations/visualizations/LinkViz/types.ts
@@ -1,0 +1,17 @@
+import { SearchModelType, CardDisplayType } from "metabase-types/api";
+
+export type LinkEntity = {
+  id: number;
+  database_id?: number;
+  model: SearchModelType;
+  name: string;
+  description?: string;
+  display?: CardDisplayType;
+};
+
+export interface LinkCardSettings {
+  link: {
+    url?: string;
+    entity?: LinkEntity;
+  };
+}
